requirements.txt
mysql_connector_python==8.0.32
requests==2.28.2
requests_html==0.10.0
tabulate==0.9.0
